KUDU-3749 Fix range key hex encoding in ksck and auto_rebalancer
Both ksck and auto-rebalancer serialize range_key_begin into a hex
string. Make the encoding consistent and safe by treating each byte as
unsigned before widening, avoiding signed-char sign extension.

+++ src/kudu/master/auto_rebalancer.cc
@@ -611,12 +611,10 @@ Status AutoRebalancerTask::BuildClusterRawInfo(
           const auto& range_key_begin = partition.begin().range_key();
 
           // Format as hex string for consistency with ksck.
-          // TODO(KUDU-3749): Align ksck + auto-rebalancer hex encoding to avoid
-          // signed-char sign extension.
           std::ostringstream ss_range_key_begin;
           for (size_t i = 0; i < range_key_begin.size(); ++i) {
             ss_range_key_begin << std::hex << std::setw(2) << std::setfill('0')
-                               << static_cast<uint16_t>(range_key_begin[i]);
+                               << static_cast<uint16_t>(static_cast<uint8_t>(range_key_begin[i]));
           }
           tablet_summary.range_key_begin = ss_range_key_begin.str();
           VLOG(2) << "Tablet " << tablet_summary.id
